Change from 64 -> 32 bit float conversion when mixing AudioSegment objects
- Slightly faster, plus we don't really need 64 bits worth of accuracy

+++ tools/audiosegment.py
@@ -44,7 +44,7 @@ class AudioSegment():
     def __mul__(self, arg):
         if isinstance(arg, AudioSegment):
 
-            self._data = self._data.astype(np.float64) # Cast to float64 incase we overflow from the sum of the signals
+            self._data = self._data.astype(np.float32) # Cast to float32 incase we overflow from the sum of the signals
 
             # data_max = np.max(self._data)
             # arg_max = np.max(self._data)
